requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
python-docx==0.8.11
pytube>=10.8.5
pytube3==9.6.4
typing-extensions==3.10.0.0
Werkzeug==2.0.1
